chore: Remove Psalm (#38)
The current version of Psalm is not compatible with PHP 8.4. We've also
been struggling in `infection/infection` with the usage of Psalm (see
https://github.com/infection/infection/pull/2565).

Whilst I don't think we should entirely discard Psalm (I think there is
still value in having a Psalm integration for infection), having both
Psalm and PHPStan enabled for a small adapter feels more crippling than
anything.

So rather than trying to update it, this PR proposes to remove it.