ci: Rename the jobs and workflows for better readability (#40)
`ci` and `CI on` makes no sense: this is a GitHub Actions workflow or
job, it is part of the CI. Instead, it is more interesting to know if
it's a unit test or an end-to-end test.

Because the job label can't be too long (it gets truncated), I renamed
`Unit tests with PHP 8.4 using pcov` to `Unit tests with PHP 8.4
(pcov)`.